SEL Update matching
