+++ config/splat.us.stsel.yaml
@@ -36,7 +36,7 @@ segments:
       - [0x176C0, rodata]
       - [0x276C0, .rodata, 2D260]
       - [0x276F0, rodata]
-      - [0x27AD0, rodata] # Update function
+      - [0x277C8, .rodata, 2EED8] # Update function
       - [0x27AE4, rodata] # func_801B2E5C, func_801B3164, func_801B3A94
       - [0x27B14, rodata] # func_801B3164, func_801B3A94
       - [0x27B4C, rodata] # func_801B3A94, TestCollision
@@ -50,6 +50,7 @@ segments:
       - [0x2BF9C, .rodata, stream]
       - [0x2C048, c, 2C048]
       - [0x2D260, c, 2D260]
+      - [0x2EED8, c, 2EED8]
       - [0x3410C, c, 3410C]
       - [0x3585C, c, 3585C]
       - [0x3642C, c, 3642C]
